% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QI_F.R
\docType{data}
\name{df}
\alias{df}
\title{Example Data Frame for Quantity-intensity relationship}
\format{
Data frame with Solution: soil ratio in the first column, Initial K concentration (mg/L) in the second column, Final or equilibrium K concentration (mg/L) in the third column, Final or equilibrium 'Ca+Mg' concentration (mol/L) in the fourth column. Write the following notations on the spreadsheet:

Solution_to_Soil_Ratio - for Solution: soil ratio

Initial_K - for Initial K concentration (mg/L)

Final_K - for Final or equilibrium K concentration (mg/L)

Final_Ca_and_Mg - for Final or equilibrium 'Ca+Mg' concentration (mol/L)
}
\usage{
df
}
\description{
User is advised to prepare the data as suggested in the example to derive Quantity-Intensity (Q/I) relationship parameters of soil potassium (K) using linear and polynomial (second order) regression
}
\keyword{datasets}
